Allow Creating New Session in Cram Mode
A current session does not exist when `org-drill-cram` is called, so
it's current behavior of toggling the `crame-mode` property doesn't
seem to do anything. This change allows it to inform `org-drill` to
create a new session in "cram mode."

I'm not sure if this is the proper way to impliment this relative to
the rest of the codebase, which I'm still a little unfamiliar
with. But I thought this pr would still be useful, if only to create discussion.